Add more basic functions to site
Create a simple homepage with new Thymeleaf template named
home.html. The template is given an attribute "shortURLFormData" which
will retrieve user input data from the form.

The data is then retrieved and a new ShortURL created, with the
shortened link returned to the user on the page, using the template
success.html.

* Add Bootstrap CSS dependency to home.html and success.html through
their CDN

* Move table of ShortURLs to /table. ShortURLs can be freely added and
deleted using this page