Define /logout requests as explicit POSTs for CSRF
By default, the given logout page from Spring Security does not contain
CSRF tokens. To solve this, hidden forms are now placed and called with
JavaScript when the logout button is clicked. This form will have the
CSRF token required for the POSt to go through, as supplied by Thymeleaf
using the th:action tag.