Create tests for restrictive owner views
Said tests have been placed in ApiAuthorizedViewTest.java . Do note that
this test class requires a POST request to the ShortURL form in the
homepage, as this is the only way to create a ShortURL anonymously (with
a undefined/null owner).

Updated tests in ApiPostRequestTest.java to use the static field
SHORTURL_API from the template.

Changes to UserAndSecurityTestTemplate.java:

- Clarified instructions on how to have test data persist for debugging.

- Added helper methods: one for adding random ShortURLs with a specified
  owner and one for getting all ShortURLs of one owner, with an argument
  to check for the expected amount.