Keep current page saved in state
When deleting entries from the table in a page other than the first,
reloading will cause the table to switch back to the first page. These
changes will switch the page to whatever was saved in the state.
Entering a new value for maximum table size will reset the saved page to
zero (0).