Change OAuth UUID field definition to provider
Instead of taking the hashcode of the entire user attributes map, the
hashcode of the map's keyset is taken instead, because the values for
the attribute map may not remain constant over time, but the keys likely
will.

The purpose of this field was to differeniate between providers to clarify
issues of duplicate IDs between users from different services. Doing it
in this manner will still serve this purpose correctly.