Update README for OAuth2 and new admin account
Due to browsers beginning to warning users of the dangers of credential
stuffing, the default admin acccount of admin:password has been changed
to relinkadmin:relinkpassword

Such a warning can cause massive confusion, so it would be best to avoid
it as much as possible. Possibly enforce password strength rules later
on to help bolster against it.